refactor: rename of bottomsheet component to index
