+++ src/components/PaymentModal/AddCardStep/AddCardStep.tsx
@@ -1,5 +1,5 @@
 import { Card } from '@/components/Card'
-import { BottomSheet } from '@/components/common/BottomSheet/BottomSheet'
+import { BottomSheet } from '@/components/common/BottomSheet'
 import { CardDateInput } from '@/components/common/CardDateInput/CardDateInput'
 import { CardNumberInput } from '@/components/common/CardNumberInput/CardNumberInput'
 import { CardPasswordInput } from '@/components/common/CardPasswordInput/CardPasswordInput'
